requirements.txt
# TODO:
torch~=1.13.0
argparse~=1.4.0
PyYAML~=6.0
torchsummary~=1.5.1
torchvision~=0.14.0